Move FileUtil.getConfiguredXStream() into PersistenceUtil instead
This lets FileUtil be more generic rather than supporting MT directly.
