+++ src/main/java/net/rptools/lib/FileUtil.java
@@ -41,10 +41,7 @@ import java.util.regex.Pattern;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
 import java.util.zip.ZipInputStream;
-import net.rptools.maptool.client.ui.token.BarTokenOverlay;
-import net.rptools.maptool.model.AStarCellPointConverter;
-import net.rptools.maptool.model.ShapeType;
-import net.rptools.maptool.model.converters.WallTopologyConverter;
+import net.rptools.maptool.util.PersistenceUtil;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
 import org.apache.logging.log4j.LogManager;
@@ -80,7 +77,7 @@ public class FileUtil {
   }
 
   public static Object objFromResource(String res) throws IOException {
-    XStream xs = getConfiguredXStream();
+    XStream xs = PersistenceUtil.getConfiguredXStream();
     try (InputStream is = FileUtil.class.getClassLoader().getResourceAsStream(res)) {
       return xs.fromXML(new InputStreamReader(is, StandardCharsets.UTF_8));
     }
@@ -479,22 +476,5 @@ public class FileUtil {
    */
   public static String stripInvalidCharacters(String fileName) {
     return fileName = fileName.replaceAll("[^\\w\\s.,-]", "_");
-  }
-
-  /**
-   * Return an XStream which allows net.rptools.**, java.awt.**, sun.awt.** May be too permissive,
-   * but it Works For Me(tm)
-   *
-   * @return a configured XStream
-   */
-  public static XStream getConfiguredXStream() {
-    XStream xStream = new XStream();
-    XStream.setupDefaultSecurity(xStream);
-    xStream.allowTypesByWildcard(new String[] {"net.rptools.**", "java.awt.**", "sun.awt.**"});
-    xStream.registerConverter(new AStarCellPointConverter());
-    xStream.registerConverter(new WallTopologyConverter(xStream));
-    xStream.addImmutableType(ShapeType.class, true);
-    xStream.addImmutableType(BarTokenOverlay.Side.class, true);
-    return xStream;
   }
 }
